Unstable update for handling integrals and diff
This update tries to extend the functionalty of derive() so it can handle integrals. However, this stil not stable, and I should puse working on this untile the derive algorithem becoms more controlable and better in tracking the deravation process.

Note: Many other things have been updated as well, and I will marge them to master but without the handling integrals function.